Cleaned up saving and loading currentUser to IndexedDB Handled everything in store now
